Move getMemberInfo test into component suite
